Fixed declaration bug

+++ src/ESP32.cpp
@@ -17,7 +17,7 @@ int ESP32::readRegister(int address){
 	return address;
 }
 
-void ESP32::writeRegister(uint8_t address, std::array<uint8_t, 2> data) {
+void ESP32::writeRegister(int address, std::array<uint8_t, 2> data) {
     auto comm = Command(1, address);
     comm.setData(data);
     comm.sendCommand();
